requirements.txt
aiowebostv==0.1.3
asyncio>=3.4.3
greenlet==3.0.1
numpy==1.24.4
opencv-python==4.8.1.78
pillow==10.4.0
playwright==1.40.0
pyee==11.0.1
pylgtv==0.1.9
python-dotenv==1.0.0
typing_extensions==4.13.2
websockets>=12.0
